pyproject.toml
[project]
name = "upbit-trader"
version = "0.1.0"
description = "Upbit 암호화폐 자동매매 시스템"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27",
    "websockets>=13.0",
    "pandas>=2.2",
    "numpy>=1.26",
    "pandas-ta>=0.3.14b1",
    "sqlalchemy>=2.0",
    "aiosqlite>=0.20",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "pyjwt>=2.8",
    "apscheduler>=3.10",
    "python-telegram-bot>=21.0",
    "greenlet>=3.3.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "ruff>=0.5",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100
